Code refactoring for bubble chart
