+++ app/frontend/utils/topic-utils.tsx
@@ -71,46 +71,6 @@ const TopicUtils = {
     }
 
     return `${attributedDelta.toLocaleString('en-US')} (${formattedPercentage})`;
-  },
-
-  fieldsForStat(stat: String) {
-    switch(stat) {
-      case 'articles':
-        return {
-          totalField: 'articles_count',
-          deltaField: 'articles_count_delta',
-          attributedDeltaField: 'attributed_articles_created_delta'
-        }
-        break;
-      case 'revisions':
-        return {
-          totalField: 'revisions_count',
-          deltaField: 'revisions_count_delta',
-          attributedDeltaField: 'attributed_revisions_count_delta'
-        }
-        break;
-      case 'length':
-        return {
-          totalField: 'length',
-          deltaField: 'length_delta',
-          attributedDeltaField: 'attributed_length_delta'
-        }
-        break;
-      case 'tokens':
-        return {
-          totalField: 'token_count',
-          deltaField: 'token_count_delta',
-          attributedDeltaField: 'attributed_token_count'
-        }
-        break;
-      case 'wp10':
-        return {
-          totalField: 'average_wp10_prediction',
-          deltaField: 'articles_count_delta',
-          attributedDeltaField: 'attributed_articles_created_delta',
-        }
-        break;
-    }
   }
 }
